requirements.txt
praw
moviepy
pydub
requests
pillow